spectra-api/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
PyJWT>=2.8.0
certifi>=2024.2.2
pg8000>=1.31.2
cryptography>=42.0.0
